Initialise components for chat page
