+++ frontend/src/components/chat_component/ChatBox.js
@@ -1,0 +1,9 @@
+
+
+const ChatBox = () => {
+    return (
+        <>ChatBox</>
+    );
+}
+
+export default ChatBox
