+++ frontend/src/components/chat_component/UserList.js
@@ -1,0 +1,9 @@
+
+
+const UserList = () => {
+    return (
+        <>UserList</>
+    );
+}
+
+export default UserList
